+++ tests/phpunit/tests/add-admin-css.php
@@ -613,7 +613,7 @@ class Add_Admin_CSS_Test extends WP_UnitTestCase {
 		$this->test_can_show_css_with_true_query_param_by_admin();
 
 		$expected = "			<div class=\"notice notice-error\">
-				<p><strong>RECOVERY MODE ENABLED:</strong> CSS output for this plugin is disabled on this page view.</p>
+				<p><strong>RECOVERY MODE ENABLED:</strong> CSS output for the Add Admin CSS plugin is disabled on this page view.</p>
 			</div>";
 
 		$this->assertEquals( $expected, $this->get_action_output( 'admin_notices' ) );
@@ -723,7 +723,7 @@ class Add_Admin_CSS_Test extends WP_UnitTestCase {
 		$this->fake_current_screen();
 
 		$expected = "			<div class=\"notice notice-error\">
-				<p><strong>RECOVERY MODE ENABLED:</strong> CSS output for this plugin is currently disabled for the entire admin area via use of the <code>C2C_ADD_ADMIN_CSS_DISABLED</code> constant.</p>
+				<p><strong>RECOVERY MODE ENABLED:</strong> CSS output for the Add Admin CSS plugin is currently disabled for the entire admin area via use of the <code>C2C_ADD_ADMIN_CSS_DISABLED</code> constant.</p>
 			</div>";
 
 		$this->assertEquals( $expected, $this->get_action_output( 'admin_notices' ) );